Attempt to fix messages

+++ db/schema.rb
@@ -16,8 +16,8 @@ ActiveRecord::Schema.define(version: 2021_03_26_083400) do
   enable_extension "plpgsql"
 
   create_table "messages", force: :cascade do |t|
-    t.integer "room_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "room_id", null: false
+    t.bigint "user_id", null: false
     t.text "body"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
@@ -27,8 +27,8 @@ ActiveRecord::Schema.define(version: 2021_03_26_083400) do
   end
 
   create_table "room_users", force: :cascade do |t|
-    t.integer "room_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "room_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.datetime "last_read_at"
